Split parse/codegen module into codegen/ package
